refactor: centralize creditor name normalization
