+++ main.py
@@ -135,7 +135,7 @@ def extract_all_accounts(sections: dict) -> list:
 
     from datetime import datetime
     import re
-    from logic.generate_goodwill_letters import normalize_creditor_name
+    from logic.utils.names_normalization import normalize_creditor_name
 
     def sanitize_number(num: str | None) -> str:
         if not num:
